+++ src/viewpanel/components/TagPicker.tsx
@@ -5,7 +5,6 @@ import { CommandToCode } from '../CommandToCode';
 import { TagType } from '../TagType';
 import { MessageHelper } from '../helper/MessageHelper';
 import Downshift from 'downshift';
-import { Icon } from './Icon';
 
 export interface ITagPickerProps {
   type: string;
@@ -15,10 +14,11 @@ export interface ITagPickerProps {
   freeform: boolean;
   focussed: boolean;
   unsetFocus: () => void;
+  disableConfigurable?: boolean;
 }
 
 export const TagPicker: React.FunctionComponent<ITagPickerProps> = (props: React.PropsWithChildren<ITagPickerProps>) => {
-  const { icon, type, crntSelected, options, freeform, focussed, unsetFocus } = props;
+  const { icon, type, crntSelected, options, freeform, focussed, unsetFocus, disableConfigurable } = props;
   const [ selected, setSelected ] = React.useState<string[]>([]);
   const [ inputValue, setInputValue ] = React.useState<string>("");
   const prevSelected = usePrevious(crntSelected);
@@ -49,7 +49,16 @@ export const TagPicker: React.FunctionComponent<ITagPickerProps> = (props: React
    * @param values 
    */
   const sendUpdate = (values: string[]) => {
-    const cmdType = type === TagType.tags ? CommandToCode.updateTags : CommandToCode.updateCategories;
+    let cmdType = CommandToCode.updateCategories;
+    
+    if (type === TagType.tags) {
+      cmdType = CommandToCode.updateTags;
+    } else if (type === TagType.categories) {
+      cmdType = CommandToCode.updateCategories;
+    } else if (type === TagType.keywords) {
+      cmdType = CommandToCode.updateKeywords;
+    }
+
     MessageHelper.sendMessage(cmdType, values);
   };
 
@@ -133,6 +142,12 @@ export const TagPicker: React.FunctionComponent<ITagPickerProps> = (props: React
                             ref: inputRef, 
                             onFocus: openMenu as any, 
                             onClick: openMenu as any,
+                            onKeyDown: (e) => {
+                              if (e.key === 'Enter') {
+                                e.preventDefault();
+                                insertUnkownTag(closeMenu);
+                              }
+                            },
                             onBlur: () => { 
                               closeMenu(); 
                               unsetFocus(); 
@@ -150,7 +165,7 @@ export const TagPicker: React.FunctionComponent<ITagPickerProps> = (props: React
                             title={`Add the unknown tag`}
                             disabled={!inputValue} 
                             onClick={() => insertUnkownTag(closeMenu)}>
-                      <Icon name="tag" />
+                      {icon}
                     </button>
                   )
                 }
@@ -170,7 +185,7 @@ export const TagPicker: React.FunctionComponent<ITagPickerProps> = (props: React
         }
       </Downshift>
 
-      <Tags values={selected.sort((a: string, b: string) => a.toLowerCase() < b.toLowerCase() ? -1 : 1 )} onRemove={onRemove} onCreate={onCreate} options={options} />
+      <Tags values={selected.sort((a: string, b: string) => a.toLowerCase() < b.toLowerCase() ? -1 : 1 )} onRemove={onRemove} onCreate={onCreate} options={options} disableConfigurable={!!disableConfigurable} />
     </div>
   );
 };